+++ imports/components/@primitives/modals/index.js
@@ -1,4 +1,3 @@
-
 import { Meteor } from "meteor/meteor";
 import { Component, PropTypes } from "react";
 import { connect } from "react-redux";
@@ -6,20 +5,19 @@ import { connect } from "react-redux";
 import { modal as modalActions, nav as navActions } from "../../../data/store";
 
 import Modal from "./SideModal";
+import PromptModal from "./PromptModal";
 
 class SideModalContainerWithoutData extends Component {
-
   static propTypes = {
     dispatch: PropTypes.func,
-    modal: PropTypes.object, // eslint-disable-line
-    navigation: PropTypes.object, // eslint-disable-line
+    modal: PropTypes.object,
+    navigation: PropTypes.object,
     path: PropTypes.string,
-  }
-
+  };
 
   state = {
     previous: null,
-  }
+  };
 
   componentDidMount() {
     if (!this.props.modal.props.keepNav && this.props.modal.visible) {
@@ -40,10 +38,7 @@ class SideModalContainerWithoutData extends Component {
     ) {
       this.props.dispatch(navActions.setLevel("MODAL"));
       this.setState({ previous: this.props.navigation.level });
-    } else if (
-      nextProps.modal.visible &&
-      nextProps.navigation.level === "DOWN"
-    ) {
+    } else if (nextProps.modal.visible && nextProps.navigation.level === "DOWN") {
       this.setState({ previous: this.props.navigation.level });
     }
 
@@ -59,7 +54,7 @@ class SideModalContainerWithoutData extends Component {
       this.props.dispatch(navActions.setLevel(previous));
     }
 
-    if (!nextProps.modal.visible && (this.props.path !== nextProps.path)) {
+    if (!nextProps.modal.visible && this.props.path !== nextProps.path) {
       this.props.dispatch(modalActions.hide());
     }
   }
@@ -69,9 +64,10 @@ class SideModalContainerWithoutData extends Component {
       const root = document.documentElement;
 
       if (!nextProps.modal.visible) {
-        root.className = root.className.split(" ").filter((className) =>
-          className !== "modal--opened"
-        ).join(" ");
+        root.className = root.className
+          .split(" ")
+          .filter((className) => className !== "modal--opened")
+          .join(" ");
       } else if (!this.props.modal.visible && nextProps.modal.visible) {
         root.className += " modal--opened";
       }
@@ -94,7 +90,7 @@ class SideModalContainerWithoutData extends Component {
 
     // up arrow
     if (keyCode === 38 && this.scrollElement) this.scrollElement.scrollTop -= 10;
-  }
+  };
 
   close = (e) => {
     const { target } = e;
@@ -103,14 +99,28 @@ class SideModalContainerWithoutData extends Component {
     if (this.props.modal.props.forceOpen) return;
 
     this.props.dispatch(modalActions.hide());
-  }
+  };
 
   captureRef = (ref) => {
     this.scrollElement = ref;
-  }
+  };
 
   render() {
     const { visible, content, props } = this.props.modal;
+    if (props && props.promptModal) {
+      return (
+        <PromptModal
+          close={this.close}
+          profileImage={"https://pbs.twimg.com/profile_images/1893823484/image.jpg"}
+          heroImage={
+            "https://s3.amazonaws.com/ns.images/newspring/collection/series_newspring/v4.sermonseries.graphics.myspace.2to1.jpg"
+          }
+          component={content}
+          visible={visible}
+          {...this.props}
+        />
+      );
+    }
     return (
       <Modal
         close={this.close}
@@ -134,8 +144,4 @@ const withRedux = connect(map);
 
 export default withRedux(SideModalContainerWithoutData);
 
-export {
-  SideModalContainerWithoutData,
-  map,
-  withRedux,
-};
+export { SideModalContainerWithoutData, map, withRedux };